+++ internal/transcription/transcription.go
@@ -24,7 +24,7 @@ const (
 	transcriptionBackendName = "parakeet"
 	defaultModel             = "nvidia/parakeet-tdt-0.6b-v2"
 	defaultDevice            = "auto"
-	defaultPrecision         = "fp32"
+	defaultPrecision         = "bf16"
 )
 
 // Config defines transcription service settings.
